initial border work
