+++ src/main/java/io/github/cardsandhuskers/survivalgames/objects/border/SkywarsBorder.java
@@ -1,0 +1,2 @@
+package io.github.cardsandhuskers.survivalgames.objects.border;public class SkywarsBorder {
+}
